Move legion/*/*.hh up
Rename init_args to task_prologue (file and class)
