+++ flecsi/execution/CMakeLists.txt
@@ -35,12 +35,12 @@ if(FLECSI_RUNTIME_MODEL STREQUAL "legion")
 
   set(execution_HEADERS
     ${execution_HEADERS}
-    legion/enactment/bind_accessors.hh
-    legion/enactment/task_wrapper.hh
-    legion/enactment/unbind_accessors.hh
+    legion/bind_accessors.hh
+    legion/task_wrapper.hh
+    legion/unbind_accessors.hh
     legion/execution_policy.hh
     legion/future.hh
-    legion/invocation/init_args.hh
+    legion/task_prologue.hh
     legion/reduction_wrapper.hh
   )
