+++ flecsi/execution/legion/task_wrapper.hh
@@ -20,12 +20,12 @@
 #if !defined(__FLECSI_PRIVATE__)
 #error Do not include this file directly!
 #else
-#include "../../task_attributes.hh"
+#include "../task_attributes.hh"
+#include "bind_accessors.hh"
 #include "flecsi/runtime/backend.hh"
 #include "flecsi/utils/function_traits.hh"
 #include "flecsi/utils/serialize.hh"
-#include <flecsi/execution/legion/enactment/bind_accessors.hh>
-#include <flecsi/execution/legion/enactment/unbind_accessors.hh>
+#include "unbind_accessors.hh"
 #include <flecsi/utils/common.hh>
 #include <flecsi/utils/flog.hh>
 #endif
